[generator] Remove `main` module and class use `generator` instead
